feat : review swiper component separate & component name edit
