+++ src/components/MainSwiper/MainPageSwiper.tsx
@@ -3,7 +3,6 @@ import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import { Navigation, Pagination } from 'swiper/modules';
 import * as S from './MainPageSwiper.styles';
-import { useState } from 'react';
 import { SwiperBookItems } from './SwiperBookItems/SwiperBookItems';
 import { useNavigate } from 'react-router-dom';
 import { useGetBestBook } from '../../querys/bookQuery';
@@ -18,9 +17,6 @@ export type SwiperItems = {
 };
 
 export function MainPageSwiper() {
-  const [active, setActive] = useState(0);
-  // const [bestBooks, setBestBooks] = useState<SwiperItems[]>([]);
-
   const navigate = useNavigate();
 
   const { data: bestBooks, isLoading } = useGetBestBook();
@@ -39,12 +35,11 @@ export function MainPageSwiper() {
             nextEl: '.swiper-button-next',
             prevEl: '.swiper-button-prev',
           }}
-          onSlideChange={(swiper) => setActive(swiper.activeIndex)}
         >
           {bestBooks.map((item, idx) => (
             <S.BookSlideContainer key={idx}>
               <S.BookSlide onClick={() => navigate(`books/${item.bookId}`)}>
-                {active === idx && <SwiperBookItems item={item} />}
+                <SwiperBookItems item={item} />
               </S.BookSlide>
             </S.BookSlideContainer>
           ))}
